fantastic_four/requirements.txt
torch
torchvision
pillow
streamlit
pyngrok
numpy
